# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EduJobs Connect API",
    "description": "Comprehensive API for the EduJobs Connect platform - connecting Rwandan students with universities and job opportunities",
    "version": "1.0.0",
    "contact": {
      "name": "EduJobs Connect Support",
      "email": "support@edujobsconnect.rw",
      "url": "https://edujobsconnect.rw"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.edujobsconnect.rw/api",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Enter your token in the text input below."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "objectId",
            "example": "507f1f77bcf86cd799439011"
          },
          "firstName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "example": "Doe"
          },
          "fullName": {
            "type": "string",
            "readOnly": true,
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john.doe@example.com"
          },
          "role": {
            "type": "string",
            "enum": ["student", "employer", "admin"],
            "default": "student",
            "example": "student"
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "suspended"],
            "default": "active",
            "example": "active"
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "profile": {
            "$ref": "#/components/schemas/UserProfile"
          },
          "preferences": {
            "$ref": "#/components/schemas/UserPreferences"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": ["firstName", "lastName", "email"]
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+250788123456"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "example": "1995-06-15"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female", "other"],
            "example": "male"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "bio": {
            "type": "string",
            "maxLength": 500,
            "example": "Computer Science student passionate about technology and innovation."
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["JavaScript", "React", "Node.js"]
          },
          "experience": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Experience"
            }
          },
          "education": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Education"
            }
          },
          "resume": {
            "type": "string",
            "format": "uri",
            "example": "/uploads/resumes/john_doe_resume.pdf"
          },
          "avatar": {
            "type": "string",
            "format": "uri",
            "example": "/uploads/avatars/john_doe.jpg"
          }
        }
      },
      "UserPreferences": {
        "type": "object",
        "properties": {
          "emailNotifications": {
            "type": "boolean",
            "default": true,
            "example": true
          },
          "smsNotifications": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "jobAlerts": {
            "type": "boolean",
            "default": true,
            "example": true
          },
          "universityAlerts": {
            "type": "boolean",
            "default": true,
            "example": true
          },
          "language": {
            "type": "string",
            "default": "en",
            "example": "en"
          },
          "timezone": {
            "type": "string",
            "default": "Africa/Kigali",
            "example": "Africa/Kigali"
          }
        }
      },
      "Address": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "example": "123 Main Street"
          },
          "city": {
            "type": "string",
            "example": "Kigali"
          },
          "province": {
            "type": "string",
            "example": "Kigali City"
          },
          "country": {
            "type": "string",
            "default": "Rwanda",
            "example": "Rwanda"
          },
          "postalCode": {
            "type": "string",
            "example": "00000"
          }
        }
      },
      "Experience": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Software Developer Intern"
          },
          "company": {
            "type": "string",
            "example": "Tech Solutions Ltd"
          },
          "location": {
            "type": "string",
            "example": "Kigali, Rwanda"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2023-06-01"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "example": "2023-08-31"
          },
          "current": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "description": {
            "type": "string",
            "example": "Developed web applications using React and Node.js"
          }
        }
      },
      "Education": {
        "type": "object",
        "properties": {
          "institution": {
            "type": "string",
            "example": "University of Rwanda"
          },
          "degree": {
            "type": "string",
            "example": "Bachelor of Computer Science"
          },
          "fieldOfStudy": {
            "type": "string",
            "example": "Computer Science"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2020-09-01"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "example": "2024-06-30"
          },
          "current": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "gpa": {
            "type": "string",
            "example": "3.8"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/users/profile": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current user profile",
        "description": "Get the authenticated user's profile information",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user profile",
        "description": "Update the authenticated user's profile information",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50
                  },
                  "lastName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50
                  },
                  "profile": {
                    "$ref": "#/components/schemas/UserProfile"
                  },
                  "preferences": {
                    "$ref": "#/components/schemas/UserPreferences"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Profile updated successfully"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}